+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`2.23.0...main`][2.23.0...main].
+For a full diff see [`2.24.0...main`][2.24.0...main].
+
+## [`2.24.0`][2.24.0]
+
+For a full diff see [`2.23.0...2.24.0`][2.23.0...2.24.0].
 
 ### Added
 
@@ -383,6 +387,7 @@ For a full diff see [`7afa59c...1.0.0`][7afa59c...1.0.0].
 [2.22.1]: https://github.com/ergebnis/phpunit-slow-test-detector/releases/tag/2.22.1
 [2.22.2]: https://github.com/ergebnis/phpunit-slow-test-detector/releases/tag/2.22.2
 [2.23.0]: https://github.com/ergebnis/phpunit-slow-test-detector/releases/tag/2.23.0
+[2.24.0]: https://github.com/ergebnis/phpunit-slow-test-detector/releases/tag/2.24.0
 
 [7afa59c...1.0.0]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/7afa59c...1.0.0
 [1.0.0...2.0.0]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/1.0.0...2.0.0
@@ -416,7 +421,8 @@ For a full diff see [`7afa59c...1.0.0`][7afa59c...1.0.0].
 [2.21.0...2.22.0]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.21.0...2.22.0
 [2.22.0...2.22.1]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.22.0...2.22.1
 [2.22.1...2.22.2]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.22.1...2.22.2
-[2.23.0...main]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.23.0...main
+[2.24.0...main]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.24.0...main
+[2.23.0...2.24.0]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.23.0...2.24.0
 [2.22.2...2.23.0]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/2.22.2...2.23.0
 
 [#6]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/6